Implement AVX512 functions (broken)
